ingest/pyproject.toml
[project]
name = "ingest"
version = "0.1.0"
description = "Embedding + ingestion script for database"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "sentence-transformers",
  "torch",
  "tqdm",
   "fastapi",
  "uvicorn[standard]",
  "sentence-transformers",
  "torch",
  "pillow",
 "requests",
]

[tool.uv]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/gpu"

